Remove deprecated documentation templates and update index exports
- Deleted unused documentation templates: `playbookTemplate.ts`, `apiReferenceTemplate.ts`, `architectureTemplate.ts`, `dataFlowTemplate.ts`, `developmentWorkflowTemplate.ts`, `glossaryTemplate.ts`, `migrationTemplate.ts`, `onboardingTemplate.ts`, `projectOverviewTemplate.ts`, `securityTemplate.ts`, `testingTemplate.ts`, and `troubleshootingTemplate.ts`.
- Updated `index.ts` to remove exports for the deleted templates, streamlining the documentation generation process.
