+++ src/generators/agents/templates/index.ts
@@ -1,3 +1,2 @@
-export { renderAgentPlaybook } from './playbookTemplate';
 export { renderAgentIndex } from './indexTemplate';
 export type { AgentTemplateContext, DocTouchpoint } from './types';
