+++ src/generators/documentation/templates/index.ts
@@ -1,14 +1,2 @@
 export { renderIndex } from './indexTemplate';
-export { renderProjectOverview } from './projectOverviewTemplate';
-export { renderArchitectureNotes } from './architectureTemplate';
-export { renderDevelopmentWorkflow } from './developmentWorkflowTemplate';
-export { renderTestingStrategy } from './testingTemplate';
-export { renderGlossary } from './glossaryTemplate';
-export { renderDataFlow } from './dataFlowTemplate';
-export { renderSecurity } from './securityTemplate';
-export { renderToolingGuide } from './toolingTemplate';
-export { renderApiReference } from './apiReferenceTemplate';
-export { renderTroubleshooting } from './troubleshootingTemplate';
-export { renderMigration } from './migrationTemplate';
-export { renderOnboarding } from './onboardingTemplate';
 export type { DocumentationTemplateContext, GuideMeta, DirectoryStat } from './types';
