--- src/generators/documentation/templates/testingTemplate.ts
@@ -1,39 +1,0 @@
-/**
- * REFERENCE ONLY - This file is not used by generators anymore.
- *
- * Scaffold structures are now defined in:
- * src/generators/shared/scaffoldStructures.ts
- *
- * This file serves as historical reference for the structure/content
- * that should be generated for this document type.
- *
- * @deprecated Since v2.0.0 scaffold system
- */
-import { wrapWithFrontMatter } from './common';
-
-export function renderTestingStrategy(): string {
-  const content = `# Testing Strategy
-
-Document how quality is maintained across the codebase.
-
-## Test Types
-- Unit: List frameworks (e.g., Jest) and file naming conventions.
-- Integration: Describe scenarios and required tooling.
-- End-to-end: Note harnesses or environments if applicable.
-
-## Running Tests
-- Execute all tests with \`npm run test\`.
-- Use watch mode locally: \`npm run test -- --watch\`.
-- Add coverage runs before releases: \`npm run test -- --coverage\`.
-
-## Quality Gates
-- Define minimum coverage expectations.
-- Capture linting or formatting requirements before merging.
-
-## Troubleshooting
-
-Document flaky suites, long-running tests, or environment quirks.
-`;
-
-  return wrapWithFrontMatter(content);
-}
