+++ src/generators/shared/index.ts
@@ -1,3 +1,6 @@
 export { GeneratorUtils } from './generatorUtils';
 export { ContextGenerator } from './contextGenerator';
 export { formatDirectoryList } from './directoryTemplateHelpers';
+
+// Scaffold structures (backward compatibility & new modular structure)
+export * from './structures';
