+++ src/generators/shared/structures/agents/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Agent structure exports
+ */
+
+export { createAgentStructure } from './factory';
+export {
+  codeReviewerStructure,
+  bugFixerStructure,
+  featureDeveloperStructure,
+  refactoringSpecialistStructure,
+  testWriterStructure,
+  documentationWriterStructure,
+  performanceOptimizerStructure,
+  securityAuditorStructure,
+  backendSpecialistStructure,
+  frontendSpecialistStructure,
+  architectSpecialistStructure,
+  devopsSpecialistStructure,
+  databaseSpecialistStructure,
+  mobileSpecialistStructure,
+} from './definitions';
