+++ src/generators/shared/structures/documentation/apiReference.ts
@@ -1,0 +1,98 @@
+import { ScaffoldStructure } from '../types';
+
+export const apiReferenceStructure: ScaffoldStructure = {
+  fileType: 'doc',
+  documentName: 'api-reference',
+  title: 'API Reference',
+  description: 'Complete API documentation for programmatic and AI agent access',
+  tone: 'technical',
+  audience: 'mixed',
+  sections: [
+    {
+      heading: 'API Reference',
+      order: 1,
+      contentType: 'prose',
+      guidance: 'Describe the API purpose and provide agent-specific instructions for programmatic interaction.',
+      exampleContent: '**Agent Usage Instructions:**\n- Use documented curl commands as templates\n- Verify expected response codes\n- Handle authentication token expiration automatically\n- Parse error responses to determine retry strategy\n- Log all API interactions for debugging',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Base URLs',
+      order: 2,
+      contentType: 'code-block',
+      guidance: 'List API base URLs for different environments (dev, staging, production) with example environment variable configuration.',
+      exampleContent: '```bash\nexport API_BASE_DEV="http://localhost:3000/api"\nexport API_BASE_STAGING="https://staging-api.example.com"\nexport API_BASE_PROD="https://api.example.com"\n```',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Authentication',
+      order: 3,
+      contentType: 'prose',
+      guidance: 'Document authentication flow including token generation, storage, expiration handling, and refresh mechanisms with curl examples for agents.',
+      exampleContent: 'Include bash examples for: 1) Obtain token via POST /auth/login, 2) Use token in requests with Bearer header, 3) Check token expiration and refresh if needed',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'API Versioning',
+      order: 4,
+      contentType: 'prose',
+      guidance: 'Document API versioning strategy (URL path, header, or query parameter), current version, and deprecation policy.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Rate Limiting',
+      order: 5,
+      contentType: 'prose',
+      guidance: 'Document rate limits with code examples showing agents how to check headers and handle backoff.',
+      exampleContent: 'Include monitoring of X-RateLimit-* headers and handling 429 responses with exponential backoff',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Common Headers',
+      order: 6,
+      contentType: 'list',
+      guidance: 'List request and response headers including Content-Type, Authorization, X-Request-ID, and any custom headers.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Endpoints',
+      order: 7,
+      contentType: 'prose',
+      guidance: 'Document API endpoints organized by resource with request/response examples for GET, POST, PUT, PATCH, DELETE methods.',
+      exampleContent: 'Include: method, path, parameters, request body examples, response codes, error handling for each endpoint',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Webhooks',
+      order: 8,
+      contentType: 'prose',
+      guidance: 'Document webhook registration, available events, payload structure, and security mechanisms.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Error Responses',
+      order: 9,
+      contentType: 'prose',
+      guidance: 'Document standard error response format, HTTP status codes, and common error codes with their meanings.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'SDKs & Client Libraries',
+      order: 10,
+      contentType: 'list',
+      guidance: 'Link to available SDKs (JavaScript, Python, etc.) and OpenAPI/Swagger specifications.',
+      required: false,
+      headingLevel: 2,
+    },
+  ],
+  linkTo: ['architecture.md', 'security.md'],
+};
