+++ src/generators/shared/structures/documentation/architecture.ts
@@ -1,0 +1,119 @@
+import { ScaffoldStructure } from '../types';
+
+export const architectureStructure: ScaffoldStructure = {
+  fileType: 'doc',
+  documentName: 'architecture',
+  title: 'Architecture Notes',
+  description: 'System architecture, layers, patterns, and design decisions',
+  tone: 'technical',
+  audience: 'architects',
+  sections: [
+    {
+      heading: 'Architecture Notes',
+      order: 1,
+      contentType: 'prose',
+      guidance: 'Describe how the system is assembled and why the current design exists.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'System Architecture Overview',
+      order: 2,
+      contentType: 'prose',
+      guidance: 'Summarize the top-level topology (monolith, modular service, microservices) and deployment model. Highlight how requests traverse the system and where control pivots between layers.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Architectural Layers',
+      order: 3,
+      contentType: 'list',
+      guidance: 'List architecture layers with their purpose and key directories. Do NOT list symbol counts inline - reference codebase-map.json for detailed analysis.',
+      exampleContent: '- **Services**: Core business logic (`src/services/`)\n- **Generators**: Content generation (`src/generators/`)\n\n> See [`codebase-map.json`](./codebase-map.json) for complete symbol counts and dependency graphs.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Detected Design Patterns',
+      order: 4,
+      contentType: 'table',
+      guidance: 'Table with Pattern, Confidence, Locations, and Description columns. Link to actual implementations.',
+      exampleContent: '| Pattern | Confidence | Locations | Description |\n|---------|------------|-----------|-------------|\n| Factory | 85% | `LLMClientFactory` | Creates LLM client instances |',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Entry Points',
+      order: 5,
+      contentType: 'list',
+      guidance: 'List entry points with markdown links to the actual files.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Public API',
+      order: 6,
+      contentType: 'table',
+      guidance: 'Table of exported symbols with Symbol, Type, and Location columns.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Internal System Boundaries',
+      order: 7,
+      contentType: 'prose',
+      guidance: 'Document seams between domains, bounded contexts, or service ownership. Note data ownership, synchronization strategies, and shared contract enforcement.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'External Service Dependencies',
+      order: 8,
+      contentType: 'list',
+      guidance: 'List SaaS platforms, third-party APIs, or infrastructure services. Describe authentication methods, rate limits, and failure considerations.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Key Decisions & Trade-offs',
+      order: 9,
+      contentType: 'prose',
+      guidance: 'Summarize architectural decisions, experiments, or ADR outcomes. Explain why selected approaches won over alternatives.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Diagrams',
+      order: 10,
+      contentType: 'diagram',
+      guidance: 'Link architectural diagrams or add mermaid definitions showing system components and their relationships.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Risks & Constraints',
+      order: 11,
+      contentType: 'prose',
+      guidance: 'Document performance constraints, scaling considerations, or external system assumptions.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Top Directories Snapshot',
+      order: 12,
+      contentType: 'list',
+      guidance: 'List top directories with approximate file counts.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Related Resources',
+      order: 13,
+      contentType: 'list',
+      guidance: 'Link to Project Overview and other relevant documentation.',
+      required: true,
+      headingLevel: 2,
+    },
+  ],
+  linkTo: ['project-overview.md', 'data-flow.md', 'codebase-map.json'],
+};
