+++ src/generators/shared/structures/documentation/glossary.ts
@@ -1,0 +1,69 @@
+import { ScaffoldStructure } from '../types';
+
+export const glossaryStructure: ScaffoldStructure = {
+  fileType: 'doc',
+  documentName: 'glossary',
+  title: 'Glossary & Domain Concepts',
+  description: 'Project terminology, type definitions, domain entities, and business rules',
+  tone: 'technical',
+  audience: 'mixed',
+  sections: [
+    {
+      heading: 'Glossary & Domain Concepts',
+      order: 1,
+      contentType: 'prose',
+      guidance: 'List project-specific terminology, acronyms, domain entities, and user personas.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Type Definitions',
+      order: 2,
+      contentType: 'list',
+      guidance: 'List exported type definitions and interfaces with links to their locations.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Enumerations',
+      order: 3,
+      contentType: 'list',
+      guidance: 'List exported enums with links to their locations.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Core Terms',
+      order: 4,
+      contentType: 'list',
+      guidance: 'Define key terms, their relevance, and where they surface in the codebase.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Acronyms & Abbreviations',
+      order: 5,
+      contentType: 'list',
+      guidance: 'Expand abbreviations and note associated services or APIs.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Personas / Actors',
+      order: 6,
+      contentType: 'prose',
+      guidance: 'Describe user goals, key workflows, and pain points addressed by the system.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Domain Rules & Invariants',
+      order: 7,
+      contentType: 'prose',
+      guidance: 'Capture business rules, validation constraints, or compliance requirements. Note region/localization nuances.',
+      required: false,
+      headingLevel: 2,
+    },
+  ],
+  linkTo: ['project-overview.md'],
+};
