+++ src/generators/shared/structures/documentation/index.ts
@@ -1,0 +1,16 @@
+/**
+ * Documentation structure definitions
+ */
+
+export { projectOverviewStructure } from './projectOverview';
+export { architectureStructure } from './architecture';
+export { developmentWorkflowStructure } from './workflow';
+export { testingStrategyStructure } from './testing';
+export { toolingStructure } from './tooling';
+export { securityStructure } from './security';
+export { glossaryStructure } from './glossary';
+export { dataFlowStructure } from './dataFlow';
+export { onboardingStructure } from './onboarding';
+export { apiReferenceStructure } from './apiReference';
+export { migrationStructure } from './migration';
+export { troubleshootingStructure } from './troubleshooting';
