+++ src/generators/shared/structures/documentation/migration.ts
@@ -1,0 +1,111 @@
+import { ScaffoldStructure } from '../types';
+
+export const migrationStructure: ScaffoldStructure = {
+  fileType: 'doc',
+  documentName: 'migration',
+  title: 'Migration Guide',
+  description: 'Complete guide for upgrading between versions, breaking changes, and data migration procedures',
+  tone: 'technical',
+  audience: 'developers',
+  sections: [
+    {
+      heading: 'Migration Guide',
+      order: 1,
+      contentType: 'prose',
+      guidance: 'Introduction explaining version support policy and the migration guide purpose.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Current Version & Support Policy',
+      order: 2,
+      contentType: 'prose',
+      guidance: 'Document latest version, release date, and version support lifecycle (how long each version receives support).',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Before You Migrate',
+      order: 3,
+      contentType: 'checklist',
+      guidance: 'Pre-migration checklist including backup, changelog review, staging test, dependency check, stakeholder notification, and rollback planning.',
+      exampleContent: '- [ ] Backup all data (database, file storage)\n- [ ] Review changelog for target version\n- [ ] Test migration in staging environment\n- [ ] Check dependency compatibility\n- [ ] Notify stakeholders\n- [ ] Plan rollback procedure\n- [ ] Schedule maintenance window',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'System Requirements',
+      order: 4,
+      contentType: 'list',
+      guidance: 'List minimum requirements for each version including Node.js, database, OS, memory, and disk space.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Migration Paths',
+      order: 5,
+      contentType: 'prose',
+      guidance: 'Document supported upgrade paths with breaking changes, step-by-step procedures with bash commands, and rollback instructions.',
+      exampleContent: 'For each major version upgrade: list breaking changes, provide detailed migration steps with code blocks, include rollback procedures',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Database Migrations',
+      order: 6,
+      contentType: 'prose',
+      guidance: 'Document database migration procedures including running migrations, creating custom migrations, and handling large datasets.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Data Transformation',
+      order: 7,
+      contentType: 'prose',
+      guidance: 'Document data transformation strategies including batch migration scripts and zero-downtime migration patterns.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Deprecation Timeline',
+      order: 8,
+      contentType: 'list',
+      guidance: 'List deprecated features and the timeline for removal with migration paths.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Backward Compatibility',
+      order: 9,
+      contentType: 'prose',
+      guidance: 'Explain compatibility mode and feature flags for managing transition between versions.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Common Issues & Solutions',
+      order: 10,
+      contentType: 'prose',
+      guidance: 'Document common migration issues and solutions including migration failures, data inconsistencies, and performance problems.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Verification & Testing',
+      order: 11,
+      contentType: 'list',
+      guidance: 'List scripts and tools for verifying migration success and testing application functionality.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Getting Help',
+      order: 12,
+      contentType: 'prose',
+      guidance: 'Provide support contact information and guidance for reporting migration issues.',
+      required: false,
+      headingLevel: 2,
+    },
+  ],
+  linkTo: ['development-workflow.md', 'architecture.md'],
+};
