+++ src/generators/shared/structures/documentation/onboarding.ts
@@ -1,0 +1,97 @@
+import { ScaffoldStructure } from '../types';
+
+export const onboardingStructure: ScaffoldStructure = {
+  fileType: 'doc',
+  documentName: 'onboarding',
+  title: 'Onboarding Guide',
+  description: 'Complete guide for getting new team members productive, including access setup and first contributions',
+  tone: 'instructional',
+  audience: 'developers',
+  sections: [
+    {
+      heading: 'Onboarding Guide',
+      order: 1,
+      contentType: 'prose',
+      guidance: 'Welcome message and high-level overview of the onboarding process.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Day 1: Access & Accounts',
+      order: 2,
+      contentType: 'checklist',
+      guidance: 'Checklist for Day 1 access setup including email, chat, code repository, issue tracking, and documentation access.',
+      exampleContent: '- [ ] Email account - Verify access to team email\n- [ ] Chat/Communication - Join Slack/Teams and relevant channels\n- [ ] Code repository - GitHub access and SSH key configured\n- [ ] Issue tracking - Jira/GitHub Issues access\n- [ ] Documentation - Wiki/Confluence access',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Day 1-2: Development Environment',
+      order: 3,
+      contentType: 'list',
+      guidance: 'Step-by-step development environment setup with code blocks for installation and configuration commands.',
+      exampleContent: '- Install required software (Node.js, npm, Git, IDE)\n- Clone repository with example bash commands\n- Configure environment variables (.env file)\n- Run database migrations if applicable\n- Verify setup with test commands',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Day 2-3: Project Understanding',
+      order: 4,
+      contentType: 'list',
+      guidance: 'Reading list of core documentation and activities for understanding the codebase, team structure, and responsibilities.',
+      exampleContent: '- [ ] Read core documentation (Project Overview, Architecture, Development Workflow, Testing Strategy)\n- [ ] Explore project structure and main entry points\n- [ ] Meet team members and understand roles\n- [ ] Review recent pull requests for code standards',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Week 1: First Contributions',
+      order: 5,
+      contentType: 'list',
+      guidance: 'Guidance for making first code contributions, learning workflow, and setting up productivity tools.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Required Access & Permissions',
+      order: 6,
+      contentType: 'list',
+      guidance: 'List of accounts, access levels, and permissions needed for different roles (repository, deployments, databases, cloud resources).',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Team Communication',
+      order: 7,
+      contentType: 'list',
+      guidance: 'Document chat channels, recurring meetings, and communication guidelines.',
+      exampleContent: '- **Channels:** #engineering, #team-name, #deployments, #incidents\n- **Meetings:** Daily standup, Sprint planning, Retrospective, Team sync\n- **Guidelines:** Use threads, mention directly when needed, default to public channels',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Learning Resources',
+      order: 8,
+      contentType: 'list',
+      guidance: 'Links to project-specific documentation, technical skill resources, company handbooks, and external resources.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Getting Help',
+      order: 9,
+      contentType: 'prose',
+      guidance: 'Escalation steps when stuck: self-debug, search resources, ask onboarding buddy, ask team channel, ask broader engineering.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Team Culture & Expectations',
+      order: 10,
+      contentType: 'prose',
+      guidance: 'Work hours, code review expectations, development principles, and growth opportunities.',
+      required: false,
+      headingLevel: 2,
+    },
+  ],
+  linkTo: ['development-workflow.md', 'project-overview.md', 'architecture.md'],
+};
