+++ src/generators/shared/structures/documentation/projectOverview.ts
@@ -1,0 +1,113 @@
+import { ScaffoldStructure } from '../types';
+
+export const projectOverviewStructure: ScaffoldStructure = {
+  fileType: 'doc',
+  documentName: 'project-overview',
+  title: 'Project Overview',
+  description: 'High-level overview of the project, its purpose, and key components',
+  tone: 'conversational',
+  audience: 'mixed',
+  sections: [
+    {
+      heading: 'Project Overview',
+      order: 1,
+      contentType: 'prose',
+      guidance: 'Summarize in 2-3 sentences what problem this project solves and who benefits from it. Focus on the value proposition.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Codebase Reference',
+      order: 2,
+      contentType: 'prose',
+      guidance: 'Add a callout pointing to codebase-map.json for detailed symbol counts, architecture layers, and dependencies. Do NOT list all symbols inline.',
+      exampleContent: '> **Detailed Analysis**: For complete symbol counts, architecture layers, and dependency graphs, see [`codebase-map.json`](./codebase-map.json).',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Quick Facts',
+      order: 3,
+      contentType: 'list',
+      guidance: 'List root directory path, primary languages (with file counts), and key entry points. Reference codebase-map.json for full analysis.',
+      exampleContent: '- Root: `/path/to/repo`\n- Languages: TypeScript (150 files), Python (20 files)\n- Entry: `src/index.ts`\n- Full analysis: [`codebase-map.json`](./codebase-map.json)',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Entry Points',
+      order: 4,
+      contentType: 'list',
+      guidance: 'List main entry points with links (CLI, server, library exports). Use markdown links with line numbers.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Key Exports',
+      order: 5,
+      contentType: 'list',
+      guidance: 'Reference codebase-map.json for the complete list.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'File Structure & Code Organization',
+      order: 6,
+      contentType: 'list',
+      guidance: 'List top-level directories with brief descriptions of their purpose.',
+      exampleContent: '- `src/` — TypeScript source files and CLI entrypoints.\n- `tests/` — Automated tests and fixtures.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Technology Stack Summary',
+      order: 7,
+      contentType: 'prose',
+      guidance: 'Outline primary runtimes, languages, and platforms in use. Note build tooling, linting, and formatting infrastructure.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Core Framework Stack',
+      order: 8,
+      contentType: 'prose',
+      guidance: 'Document core frameworks per layer (backend, frontend, data, messaging). Mention architectural patterns enforced by these frameworks.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'UI & Interaction Libraries',
+      order: 9,
+      contentType: 'prose',
+      guidance: 'List UI kits, CLI interaction helpers, or design system dependencies. Note theming, accessibility, or localization considerations.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Development Tools Overview',
+      order: 10,
+      contentType: 'prose',
+      guidance: 'Highlight essential CLIs, scripts, or developer environments. Link to Tooling guide for deeper setup.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Getting Started Checklist',
+      order: 11,
+      contentType: 'checklist',
+      guidance: 'Provide numbered steps to get a new developer productive. Include install, run, and verify steps.',
+      exampleContent: '1. Install dependencies with `npm install`.\n2. Explore the CLI by running `npm run dev`.\n3. Review Development Workflow for day-to-day tasks.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Next Steps',
+      order: 12,
+      contentType: 'prose',
+      guidance: 'Capture product positioning, key stakeholders, and links to external documentation or product specs.',
+      required: false,
+      headingLevel: 2,
+    },
+  ],
+  linkTo: ['architecture.md', 'development-workflow.md', 'tooling.md', 'codebase-map.json'],
+};
