+++ src/generators/shared/structures/documentation/troubleshooting.ts
@@ -1,0 +1,89 @@
+import { ScaffoldStructure } from '../types';
+
+export const troubleshootingStructure: ScaffoldStructure = {
+  fileType: 'doc',
+  documentName: 'troubleshooting',
+  title: 'Troubleshooting Guide',
+  description: 'Diagnostic procedures and solutions for common issues, designed for both developers and AI agents',
+  tone: 'technical',
+  audience: 'mixed',
+  sections: [
+    {
+      heading: 'Troubleshooting Guide',
+      order: 1,
+      contentType: 'prose',
+      guidance: 'Describe the purpose and agent protocol: run diagnostics, match symptoms to patterns, resolve if confident, escalate if uncertain.',
+      exampleContent: '**Agent Protocol:**\n1. Run diagnostic script to gather facts\n2. Match symptoms to known patterns\n3. Execute resolution if confidence high (>80%)\n4. Escalate to human if uncertain\n5. Log all diagnostic steps',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Agent Diagnostic Script',
+      order: 2,
+      contentType: 'code-block',
+      guidance: 'Provide bash script for agents to diagnose system state including application status, errors, resources, services, and processes.',
+      exampleContent: 'Script that checks: health endpoint, recent errors, CPU/memory/disk usage, database/cache connectivity, recent changes, and processes',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Agent Decision Tree',
+      order: 3,
+      contentType: 'diagram',
+      guidance: 'ASCII or Mermaid diagram showing how agents classify issues based on diagnostic output (not responding, database unreachable, errors, performance).',
+      exampleContent: 'App responding? → DB reachable? → Error log entries? → Resource usage normal?',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Common Issues & Solutions',
+      order: 4,
+      contentType: 'prose',
+      guidance: 'Document common issues organized by category (startup, database, performance) with: error detection patterns, agent diagnostic procedures, auto-resolution scripts, and escalation criteria.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Application Won\'t Start',
+      order: 5,
+      contentType: 'list',
+      guidance: 'List common startup issues (port in use, missing modules, missing environment variables) with bash commands for diagnosis and resolution.',
+      exampleContent: '- Port in use: use lsof to find process, kill it, restart app\n- Missing module: clear node_modules and reinstall\n- Missing env var: check .env.example and set required variables',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Performance Issues',
+      order: 6,
+      contentType: 'prose',
+      guidance: 'Document how to diagnose performance problems (high CPU, memory, slow queries) with monitoring commands and optimization steps.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Database Issues',
+      order: 7,
+      contentType: 'prose',
+      guidance: 'Document database connection diagnostics with connection checking procedures and startup attempts for containerized or local databases.',
+      required: false,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Escalation Criteria',
+      order: 8,
+      contentType: 'prose',
+      guidance: 'Define when agents must escalate: unknown patterns, data corruption risk, security issues, multiple failed attempts, or human approval needed.',
+      required: true,
+      headingLevel: 2,
+    },
+    {
+      heading: 'Escalation Format',
+      order: 9,
+      contentType: 'code-block',
+      guidance: 'Provide template for structured escalation including issue description, symptoms, diagnostics run, attempted resolutions, and recommendations.',
+      required: false,
+      headingLevel: 2,
+    },
+  ],
+  linkTo: ['development-workflow.md', 'security.md'],
+};
