+++ src/generators/shared/structures/index.ts
@@ -1,0 +1,26 @@
+/**
+ * Scaffold structures index
+ *
+ * This file serves as the main barrel export for all scaffold structures,
+ * providing backward compatibility with the original scaffoldStructures.ts import path.
+ */
+
+// Types
+export * from './types';
+
+// Documentation structures
+export * from './documentation';
+
+// Agent structures
+export * from './agents';
+
+// Skill structures
+export * from './skills';
+
+// Plan structures
+export * from './plans';
+
+// Registry & utilities
+export * from './registry';
+export * from './validation';
+export * from './serialization';
