+++ src/generators/shared/structures/plans/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Plan structure exports
+ */
+
+export { planStructure } from './planStructure';
