+++ src/generators/shared/structures/skills/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Skill structure exports
+ */
+
+export { createSkillStructure } from './factory';
+export {
+  commitMessageSkillStructure,
+  prReviewSkillStructure,
+  codeReviewSkillStructure,
+  testGenerationSkillStructure,
+  documentationSkillStructure,
+  refactoringSkillStructure,
+  bugInvestigationSkillStructure,
+  featureBreakdownSkillStructure,
+  apiDesignSkillStructure,
+  securityAuditSkillStructure,
+} from './definitions';
